Add highlight and autopick text to hovered and autopicked cards
When hovered over, cards now turn slightly brighter. Additionally, the
user can click a card to set it to be autopicked. The card then stays
highlighted, and has the text 'autopick' on its bottom border. If the
user picks another card, the old card is returned to normal and the new
assumes autopick behavior.