Implement autopicking
When the user selects a card for autopicking, but then their time
expires, the server automatically picks the autopick card on their
behalf rather than picking a card at random. This is behavior similar
to that of MTGO.

If the user fails to select any card at all, the old behavior applies
and one is selected at random.